Introducing IUserInterfaceManager
