+++ TabloidCLI/IUserInterfaceManager.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace TabloidCLI
+{
+    public interface IUserInterfaceManager
+    {
+        IUserInterfaceManager Execute();
+    }
+}
